+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Features
 
-- Добавлена поддержка передачи JWT в куках
+- Добавлена поддержка передачи JWT в httpOnly cookies
 
 ## [0.3.0](https://github.com/steroids/nest-auth/compare/0.2.3...0.3.0) (2025-12-26)